requirements.txt
asgiref==3.8.1
Django==4.2.11
django-rest-auth==0.9.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
six==1.16.0
sqlparse==0.4.4
